refactor: add missing file extensions in import statements
