+++ components/text-area/src/text-area-field/text-area-field.js
@@ -3,7 +3,7 @@ import { Box } from '@dhis2-ui/box'
 import { Field } from '@dhis2-ui/field'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { TextArea } from '../text-area'
+import { TextArea } from '../text-area/index.js'
 
 const TextAreaField = ({
     className,
